+++ symbols.c
@@ -39,7 +39,7 @@ int find_symbol_address(struct kernsym *sym, const char *symbol_name) {
 
 	sym->size = sym->end_addr - sym->addr;
 	sym->new_size = sym->size;
-	sym->run = (unsigned long) sym->addr;
+	sym->run = sym->addr;
 
 	return 0;
 }
@@ -207,7 +207,7 @@ int find_symbol_address(struct kernsym *sym, const char *symbol_name) {
 		kfree(filename);
 	}
 
-	sym->run = (unsigned long) sym->addr;
+	sym->run = sym->addr;
 
 	out:
